pyproject.toml
[tool.poetry]
name = "myelin-sync-labels"
version = "0.1.0"
description = ""
authors = ["Ruben Schmidmeister <ruben.schmidmeister@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyGithub = "^1.43"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
